Fix some broken doc links
