+++ examples/wasm-example/src/routes/signing.rs
@@ -163,7 +163,7 @@ impl Component for SigningExamplesComponent {
                                 let signed_extrinsic = partial_signed.sign_with_address_and_signature(&account_id.into(), &multi_signature);
 
                                 // do a dry run (to debug in the js console if the extrinsic would work)
-                                let dry_res = signed_extrinsic.dry_run(None).await;
+                                let dry_res = signed_extrinsic.dry_run().await;
                                 web_sys::console::log_1(&format!("Dry Run Result: {:?}", dry_res).into());
 
                                 // return the signature and signed extrinsic
